chore: move some files / rename folder
