+++ components/delegations/IncomingDelegationsTable.tsx
@@ -3,7 +3,7 @@
 import { getCoreRowModel, useReactTable } from '@tanstack/react-table';
 import Card from 'components/common/Card';
 import Table from 'components/common/table/Table';
-import type { Delegation } from 'lib/delegate/DelegatePlatform';
+import type { Delegation } from 'lib/delegations/DelegatePlatform';
 import { useTranslations } from 'next-intl';
 import { incomingColumns } from './columns';
