+++ components/delegations/cells/ControlsCell.tsx
@@ -1,7 +1,7 @@
 import ControlsWrapper from 'components/allowances/controls/ControlsWrapper';
 import RevokeButton from 'components/allowances/controls/RevokeButton';
-import type { Delegation } from 'lib/delegate/DelegatePlatform';
-import { getDelegationKey, useRevokeDelegation } from 'lib/hooks/ethereum/useRevokeDelegation';
+import type { Delegation } from 'lib/delegations/DelegatePlatform';
+import { getDelegationKey, useRevokeDelegation } from 'lib/hooks/ethereum/delegations/useRevokeDelegation';
 
 interface Props {
   delegation: Delegation;
