+++ components/delegations/cells/DelegatorCell.tsx
@@ -1,6 +1,6 @@
 'use client';
 import AddressCell from 'components/allowances/dashboard/cells/AddressCell';
-import type { Delegation } from 'lib/delegate/DelegatePlatform';
+import type { Delegation } from 'lib/delegations/DelegatePlatform';
 
 interface Props {
   delegation: Delegation;
