+++ lib/delegations/AbstractDelegatePlatform.ts
