+++ lib/delegations/AggregateDelegatePlatform.ts
